first upload of task: 1. List length

+++ singly_linked_lists/0-print_list.c
@@ -10,7 +10,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t nodes_number = 0;
+	int nodes_number = 0;
 
 	while (h != NULL)
 	{
